requirements.txt
django>=2.2
scrapy>=1.6.0
scrapy-djangoitem>=1.1.1
scrapyd>=1.2.0
python-scrapyd-api>=2.1.2